Moved CustomerSelected to the customer list
