+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/creation/OrderCreationCustomerAddFragment.kt
@@ -125,12 +125,6 @@ class OrderCreationCustomerAddFragment : BaseFragment(R.layout.fragment_creation
                     findNavController().navigateUp()
                 }
                 is AddressViewModel.SearchCustomers -> showCustomerSearchScreen()
-                is AddressViewModel.CustomerSelected -> {
-                    sharedViewModel.onCustomerAddressEdited(
-                        billingAddress = event.billingAddress,
-                        shippingAddress = event.shippingAddress
-                    )
-                }
             }
         }
     }
